Break up snapshot for state test
