+++ lib/libs/email/preview/InitialSubmissions/State/SnapshotTest/APPKSnapshot.test.tsx
@@ -1,0 +1,12 @@
+import { describe, it, expect } from "vitest";
+import { render } from "@testing-library/react";
+
+import AppKCMSEmailPreview from "../AppK";
+
+describe("Initial Submission State Email Snapshot Test", () => {
+  it("renders a AppKCMSEmailPreview Preview Template", () => {
+    const { asFragment } = render(<AppKCMSEmailPreview />);
+
+    expect(asFragment()).toMatchSnapshot();
+  });
+});
