+++ lib/libs/email/preview/InitialSubmissions/State/SnapshotTest/CHIP_SPA.test.tsx
@@ -1,0 +1,12 @@
+import { describe, it, expect } from "vitest";
+import { render } from "@testing-library/react";
+
+import ChipSpaStateEmailPreview from "../CHIP_SPA";
+
+describe("Initial Submission State Email Snapshot Test", () => {
+  it("renders a Chipspa Preview Template", () => {
+    const { asFragment } = render(<ChipSpaStateEmailPreview />);
+
+    expect(asFragment()).toMatchSnapshot();
+  });
+});
